app/requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-python-headless==4.5.2.54
Pillow==8.2.0
Werkzeug==2.0.1
